# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CodeVault API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface DashboardStats {
  totalProjects: number;
  activeDeployments: number;
  aiTokensUsed: number;
  totalRevenue: number;
  projectsThisMonth: number;
  deploymentsToday: number;
  revenueChange?: number;
  deploymentSuccessRate?: number;
}

export interface ActivityItem {
  id: number;
  type: string;
  description: string;
  /** @nullable */
  projectName?: string | null;
  createdAt: string;
}

export type AiUsageStatsDailyUsageItem = {
  date: string;
  tokens: number;
};

export interface AiUsageStats {
  totalTokens: number;
  tokensThisMonth: number;
  requestsToday: number;
  dailyUsage: AiUsageStatsDailyUsageItem[];
}

export type RevenueStatsMonthlyDataItem = {
  month: string;
  revenue: number;
};

export interface RevenueStats {
  totalRevenue: number;
  monthlyRevenue: number;
  activeSubscriptions: number;
  monthlyData: RevenueStatsMonthlyDataItem[];
}

export interface Project {
  id: number;
  name: string;
  /** @nullable */
  description: string | null;
  status: string;
  /** @nullable */
  language: string | null;
  stars: number;
  forks: number;
  isPrivate?: boolean;
  /** @nullable */
  deploymentUrl?: string | null;
  /** @nullable */
  githubUrl?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface ProjectInput {
  name: string;
  description?: string;
  language?: string;
  isPrivate?: boolean;
  /** @nullable */
  templateId?: number | null;
  /** @nullable */
  githubUrl?: string | null;
}

export interface ProjectUpdate {
  name?: string;
  description?: string;
  status?: string;
  isPrivate?: boolean;
}

export interface ProjectTemplate {
  id: number;
  name: string;
  description: string;
  language: string;
  category: string;
  stars?: number;
}

export interface AiSession {
  id: number;
  title: string;
  /** @nullable */
  projectId: number | null;
  messageCount: number;
  createdAt: string;
  updatedAt: string;
}

export interface AiSessionInput {
  title: string;
  /** @nullable */
  projectId?: number | null;
}

export interface AiMessage {
  id: number;
  sessionId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface AiMessageInput {
  content: string;
}

export interface RepoFile {
  name: string;
  path: string;
  type: string;
  size: number;
  /** @nullable */
  lastCommit?: string | null;
  /** @nullable */
  lastCommitDate?: string | null;
}

export interface Commit {
  id: number;
  sha: string;
  message: string;
  author: string;
  date: string;
  additions?: number;
  deletions?: number;
}

export interface Branch {
  name: string;
  isDefault: boolean;
  lastCommit: string;
  updatedAt: string;
  aheadBy?: number;
  behindBy?: number;
}

export interface Issue {
  id: number;
  projectId: number;
  title: string;
  /** @nullable */
  description?: string | null;
  status: string;
  priority: string;
  labels?: string[];
  createdAt: string;
}

export interface IssueInput {
  title: string;
  description?: string;
  priority?: string;
  labels?: string[];
}

export interface Deployment {
  id: number;
  projectId: number;
  projectName: string;
  status: string;
  environment: string;
  provider: string;
  /** @nullable */
  url?: string | null;
  /** @nullable */
  duration?: number | null;
  createdAt: string;
}

export interface DeploymentInput {
  projectId: number;
  environment: string;
  provider: string;
  branch?: string;
}

export interface DeploymentLog {
  id: number;
  deploymentId: number;
  level: string;
  message: string;
  timestamp: string;
}

export interface MarketplaceItem {
  id: number;
  title: string;
  description: string;
  category: string;
  price: number;
  author: string;
  downloads: number;
  rating: number;
  tags?: string[];
  /** @nullable */
  previewUrl?: string | null;
  createdAt: string;
}

export interface MarketplaceItemInput {
  title: string;
  description: string;
  category: string;
  price: number;
  tags?: string[];
}

export interface UserProfile {
  id: number;
  username: string;
  displayName: string;
  /** @nullable */
  bio: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  website?: string | null;
  totalProjects: number;
  totalStars: number;
  followers: number;
  following: number;
  achievements?: string[];
  joinedAt: string;
}

export interface ContributionDay {
  date: string;
  count: number;
  level: number;
}

export interface TradingSignal {
  id: number;
  pair: string;
  direction: string;
  entryPrice: number;
  targetPrice: number;
  stopLoss: number;
  status: string;
  confidence: number;
  /** @nullable */
  pnl?: number | null;
  createdAt: string;
}

export interface TradingSignalInput {
  pair: string;
  direction: string;
  entryPrice: number;
  targetPrice: number;
  stopLoss: number;
  confidence: number;
}

export interface TradingStats {
  totalSignals: number;
  winRate: number;
  activeSubscribers: number;
  monthlyRevenue: number;
  avgProfit: number;
  totalPnl: number;
}

export interface Subscriber {
  id: number;
  phone: string;
  name: string;
  plan: string;
  status: string;
  joinedAt: string;
}

export interface PerformanceDataPoint {
  date: string;
  winRate: number;
  signals: number;
  pnl: number;
}