# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: CodeVault API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List all projects
      responses:
        "200":
          description: Projects list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a new project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Created project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get project by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectUpdate"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /ai/sessions:
    get:
      operationId: listAiSessions
      tags: [ai]
      summary: List AI chat sessions
      responses:
        "200":
          description: Sessions list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AiSession"
    post:
      operationId: createAiSession
      tags: [ai]
      summary: Create a new AI session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiSessionInput"
      responses:
        "201":
          description: Created session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiSession"
  /ai/sessions/{id}/messages:
    get:
      operationId: getAiMessages
      tags: [ai]
      summary: Get messages for a session
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AiMessage"
    post:
      operationId: sendAiMessage
      tags: [ai]
      summary: Send a message to AI
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiMessageInput"
      responses:
        "201":
          description: AI response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiMessage"

  # Repository
...
  /deployments:
    get:
      operationId: listDeployments
      tags: [deployments]
      summary: List all deployments
      responses:
        "200":
          description: Deployments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Deployment"
    post:
      operationId: createDeployment
      tags: [deployments]
      summary: Trigger a new deployment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeploymentInput"
      responses:
        "201":
          description: Created deployment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deployment"
  /deployments/{id}:
    get:
      operationId: getDeployment
      tags: [deployments]
      summary: Get deployment by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deployment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deployment"
  /deployments/{id}/logs:
    get:
      operationId: getDeploymentLogs
      tags: [deployments]
      summary: Get deployment logs
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DeploymentLog"

  # Marketplace
  /marketplace/items:
    get:
      operationId: listMarketplaceItems
      tags: [marketplace]
      summary: List marketplace items
      responses:
        "200":
          description: Marketplace items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MarketplaceItem"
    post:
      operationId: createMarketplaceItem
      tags: [marketplace]
      summary: List an item in the marketplace
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MarketplaceItemInput"
      responses:
        "201":
          description: Created item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MarketplaceItem"
  /marketplace/items/{id}:
    get:
      operationId: getMarketplaceItem
      tags: [marketplace]
      summary: Get marketplace item by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MarketplaceItem"

  # User Profile
...
components:
  schemas:
...
    Project:
      type: object
      required: [id, name, description, status, language, stars, forks, createdAt, updatedAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
        language:
          type: ["string", "null"]
        stars:
          type: integer
        forks:
          type: integer
        isPrivate:
          type: boolean
        deploymentUrl:
          type: ["string", "null"]
        githubUrl:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    ProjectInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
        language:
          type: string
        isPrivate:
          type: boolean
        templateId:
          type: ["integer", "null"]
        githubUrl:
          type: ["string", "null"]
    ProjectUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        status:
          type: string
        isPrivate:
          type: boolean
...
    AiSession:
      type: object
      required: [id, title, projectId, messageCount, createdAt, updatedAt]
      properties:
        id:
          type: integer
        title:
          type: string
        projectId:
          type: ["integer", "null"]
        messageCount:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
    AiSessionInput:
      type: object
      required: [title]
      properties:
        title:
          type: string
        projectId:
          type: ["integer", "null"]
    AiMessage:
      type: object
      required: [id, sessionId, role, content, createdAt]
      properties:
        id:
          type: integer
        sessionId:
          type: integer
        role:
          type: string
        content:
          type: string
        createdAt:
          type: string
    AiMessageInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
...
    Deployment:
      type: object
      required: [id, projectId, projectName, status, environment, provider, createdAt]
      properties:
        id:
          type: integer
        projectId:
          type: integer
        projectName:
          type: string
        status:
          type: string
        environment:
          type: string
        provider:
          type: string
        url:
          type: ["string", "null"]
        duration:
          type: ["integer", "null"]
        createdAt:
          type: string
    DeploymentInput:
      type: object
      required: [projectId, environment, provider]
      properties:
        projectId:
          type: integer
        environment:
          type: string
        provider:
          type: string
        branch:
          type: string
    DeploymentLog:
      type: object
      required: [id, deploymentId, level, message, timestamp]
      properties:
        id:
          type: integer
        deploymentId:
          type: integer
        level:
          type: string
        message:
          type: string
        timestamp:
          type: string
    MarketplaceItem:
      type: object
      required: [id, title, description, category, price, author, downloads, rating, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        category:
          type: string
        price:
          type: number
        author:
          type: string
        downloads:
          type: integer
        rating:
          type: number
        tags:
          type: array
          items:
            type: string
        previewUrl:
          type: ["string", "null"]
        createdAt:
          type: string
    MarketplaceItemInput:
      type: object
      required: [title, description, category, price]
      properties:
        title:
          type: string
        description:
          type: string
        category:
          type: string
        price:
          type: number
        tags:
          type: array
          items:
            type: string
...