# lib/api-zod/src/generated/types/aiSession.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CodeVault API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AiSession {
  id: number;
  title: string;
  /** @nullable */
  projectId: number | null;
  messageCount: number;
  createdAt: string;
  updatedAt: string;
}