# lib/api-zod/src/generated/types/aiSessionInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CodeVault API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AiSessionInput {
  title: string;
  /** @nullable */
  projectId?: number | null;
}