# lib/api-zod/src/generated/types/deploymentInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CodeVault API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DeploymentInput {
  projectId: number;
  environment: string;
  provider: string;
  branch?: string;
}