# lib/api-zod/src/generated/types/marketplaceItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CodeVault API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MarketplaceItem {
  id: number;
  title: string;
  description: string;
  category: string;
  price: number;
  author: string;
  downloads: number;
  rating: number;
  tags?: string[];
  /** @nullable */
  previewUrl?: string | null;
  createdAt: string;
}